Refactor away from Rails-style inner modules
The original implementation was clumsy in that you had to define one or
two inner modules to route methods between instance methods and class
methods. I wasn't a big fan of that because it encourages the imprecise
naming of "including" a module also extending the module.

This implementation is more in-line with standard Ruby practice. You can
either include or extend a freedom patch and get the benefits of the
freedom patch checking.

I'd like to restore the ability to extend a `ClassMethods` module as
part of the included hook, but as an opt-in decision instead of the
default.