Reorganize CI tasks and dependencies
In order to speed up CI, this limits the dependencies that are installed
and toggles the tasks that are run to be exactly what we want to run in
CI, yet still allow the default workflow to run all of the tasks that we
want.

As part of this, we now use the CodeClimate test reporter rather than
the deprecated gem.